+++ src/features/app/api.ts
@@ -5,6 +5,7 @@
 
 import { HttpClient } from '../../core/network/httpClient.js';
 import { readAppCache, saveAppCache, AppCacheInfo } from '../../core/utils/cache.js';
+import { contextResolver } from '../../core/utils/contextResolver.js';
 
 /**
  * Craft/App item in developer list